Move helper functions to utils
All functions related to fields moved to utils/fields.py
Updated unwanted fields for categories
